chore: organize types index
